(GH-1688) Add BoltSpec helper to load bolt constructs
This creates a new helper module `BoltSpec::BoltContext` which provides
the `in_bolt_context` method. This method makes Bolt datatypes and
functions available to a block allowing code that references Bolt
constructs to be tested. The Plans module was already doing all of this
work, creating a mock executor, stubbing run_* functions, and setting
default config, inventory, and modulepath values. This moves all of that
logic into a common BoltContext module so that it can be generally
useful instead of being coupled with testing Bolt plans. This allows
things like Puppet functions that refer to the `Boltlib::TargetSpec`
datatype to be tested without having to wrap them in a Bolt plan for
testing.

Closes #1688

!feature

* **Add Boltspec helper to load Bolt constructs** ([#1688](https://github.com/puppetlabs/bolt/issues/1688))

  A new helper function `in_bolt_context` can be used to wrap code that
  references Bolt constructs, such as the Boltlib::TargetSpec datatype.
