+++ bolt_spec_spec/functions/with_datatype.pp
@@ -1,0 +1,6 @@
+function bolt_spec_spec::with_datatype (
+  Boltlib::TargetSpec $target
+) {
+  out::message("Loaded TargetSpec ${$target}")
+  $target
+}
