+++ bolt_spec_spec/spec/spec_helper.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+$LOAD_PATH.unshift File.join(__dir__, '..', '..', 'lib')
